V2 Class Car was transfered into new Java class file Car.java
